requirements.txt
# ================================
# Python 后端项目依赖包
# ================================
# 安装命令：pip install -r requirements.txt
# 如果下载慢，使用国内镜像：
#   pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/
#
# 注意：固定版本号（==）确保团队环境一致
#       如遇 Python 3.14+ 编译问题，改用 >= 安装最新兼容版
# ================================

# FastAPI - 现代、高性能的 Web 框架，用于构建 API
# 类似前端的 Express.js，专为 Python 设计
# 文档: https://fastapi.tiangolo.com
fastapi>=0.115.5

# uvicorn - ASGI 服务器，用于运行 FastAPI 应用
# 相当于 Node.js 中的 http 服务器（[standard] 包含热重载等附加功能）
uvicorn[standard]>=0.32.1

# python-dotenv - 读取 .env 环境变量文件
# 用于管理 API Key、数据库连接等敏感配置，避免把密钥写死在代码里
python-dotenv>=1.0.1

# anthropic - Anthropic 官方 SDK，用于调用 Claude 大模型
# 文档: https://docs.anthropic.com
anthropic>=0.40.0

# pydantic - 数据验证和序列化库（FastAPI 内置使用）
# 相当于 TypeScript 的类型系统，可以定义数据结构并自动校验
pydantic>=2.10.4

# httpx - 现代 HTTP 客户端，支持同步和异步请求
# 用于在后端调用其他外部 API（比 requests 更现代）
httpx>=0.28.1

# python-multipart - 处理表单数据和文件上传
python-multipart>=0.0.20

# ================================
# 数据库相关包
# ================================

# SQLAlchemy - Python 最流行的 ORM 框架
# 用 Python 类定义表结构，自动生成 SQL，不用手写 SQL
# 文档: https://docs.sqlalchemy.org
sqlalchemy>=2.0.0

# asyncpg - PostgreSQL 的异步驱动（连接器）
# SQLAlchemy 需要通过驱动才能连接 PostgreSQL
# asyncpg 是异步版本，配合 FastAPI 的 async/await 使用
asyncpg>=0.29.0

# psycopg2-binary - PostgreSQL 的同步驱动
# Alembic 迁移工具不支持异步，需要同步驱动
# binary 版本已预编译，不需要本机安装 PostgreSQL 开发库
psycopg2-binary>=2.9.0

# alembic - 数据库迁移工具
# 管理数据库表结构的版本变更（相当于数据库的 git）
# 常用命令：
#   alembic revision --autogenerate -m "说明"  生成迁移文件
#   alembic upgrade head                        应用迁移到数据库
alembic>=1.13.0
